requirements.txt
assemblyai
python-dotenv
psycopg2-binary
openai
cloudinary
Pillow 